+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ======
 
 *   (feature) Add `SimpleEntitySearchHelper::applyJsonSearch()`.
+*   (improvement) Also automatically enable `JSON_SEARCH()` MySQL function in Doctrine.
 
 
 7.12.1